Add __version__ to __init__.py upon installing
This customises the distutils build command to sneak a __version__ attribute
into a package's top-level __init__.py before installing it. The advantage
is that the version number does not get checked into the repo but remains
in the code itself. It is also hard-coded based on the version of the code
when the package is built, which is more reliable than depending on
pkg_resources matching up the correct package. Additionally, there is only
a build dependency on katversion and not a runtime dependency, so that wheels
can be safely built without requiring installations of katversion everywhere.

In cases where a package is imported locally or via a "setup.py develop"
(aka the "-e" option), katversion.get_version can still be used to get the
current version of the package. This bit of code could be added to the
__init__.py of other packages too (once again with an optional dependency
on katversion).

A side benefit is that the setup.py of other packages can simply import
the setup() of katversion, which removes the need for explicit setuptools
imports while dealing with the versioning too.