Use the next version for development versions
This aligns katversion with PEP440, which states that "developmental releases are ordered by their numerical component, immediately before the corresponding release." So 0.2 is followed by 0.3.dev10 and finally 0.3. To guess the next
release, simply increment the last numerical part (most minor bit) of the
version. This ensures that the eventual 0.3 release can be pip-installed on top
of the 0.3.dev version without removing it first.

There are still some PEP440 deviations: katversion tolerates version strings
with a text prefix like 'v1.2' and also does not stick to the strict
alphanumeric+'.' character set allowed for the local part of the version string
(the part after the '+'), which allows arbitrary branch names and a -dirty tag.
This can be handled with version normalisation though.