Automatically determine whether it is a release
It is a release version if we are on top of a commit with a tag and the repo
is clean (i.e. without uncommitted changes). Forcing the status of a release
version externally is basically lying, as release versions should only be used
for tagged commits anyway. Remove all release settings.