Override setuptools/distutils sdist command
The source distribution will be uploaded to PyPI but this currently does not
have the version hard-baked into it. Therefore override the sdist command
in the same way as the build command. Factor out the common patching part.

Some of this is inspired by python-versioneer, which alerted me to the fact
that setuptools sdist != distutils sdist and you have to check for them.