Standardise git commit hash to start with 'g'
When 'git describe' finds a tag, the commit hash in the output string
gets a 'g' prefix to indicate that the commit is from a git repository.
However, if the repo has no tags and 'git describe' is forced to come
up with something, it returns the commit hash without a 'g' in front.
In this case, add the 'g' to make the version string more consistent.
(It is also OK to remove the 'g' in the original case since we will
always use git...)