Rework git version to use valid version tags only
The previous iteration accepted any tag as the version, leading to some
nonsensical version numbers (as normalisation is only done loosely).

Instead, replace the use of 'git describe' with a 'git log' incantation
that finds all tags in the current branch. This allows the routine to
search for the latest valid version tag (and use 0.0 if none was found).
This decouples the use of tags for releases from other potential use cases
and also enforces normalisation in the main version number.

In the absence of 'git describe', use 'git status --porcelain' to determine
dirtyness. There is no need for 'git rev-list' as 'git log' subsumes this
('rev-list' could also have done a 'log' but is slightly more cumbersome).
Since we are checking version number validity in get_git_version(), the
functionality of _next_version() is now included here for convenience.