Fix is_git check
The repo_dir variable is unset in case of a RuntimeError. To keep things
simple, assume it is not a git repo if any stderr is produced, so that a
RuntimeError leads to False. This is slightly different from the original
logic, which ignored stderr and only checked stdout. The new logic means
that any stderr output leads to stdout being ignored. It stills seems
correct, though, based on the "git rev-parse --git-dir" documentation.