Support versioning of pip sdist installs
When pip installs a source tarball or 'sdist' it actually runs setup.py again,
which then attempts to get the package version from scratch, even though it
is already baked into the package. Since the SCM is not around, things go wrong.

This uses the pkginfo external package to parse the sdist version and use that
instead in this use case. The pkginfo dependency needs special handling as it
is a "pre-setup" requirement for katversion.