Rework PEP 440 normalisation via setuptools
In a sorry case of cargo culting introduced in 8d63b32 (thanks for
nothing, setuptools_scm!), pkg_resources.SetuptoolsVersion was imported
but not used, ostensibly to detect setuptools 12, which ostensibly
changed the output of pkg_resources.parse_version from a tuple to a
Version class. This broke on the latest setuptools 39, which dropped
SetuptoolsVersion in favour of the Version classes in the packaging
module.

The truth is that setuptools 8 already introduced SetuptoolsVersion
and PEP 440 parsing. Rather check directly whether parse_version
returns a tuple, in which case we fall back to the homegrown
normalisation.

This fixes katversion to work correctly with setuptools 39.