Flake8 cleanup
Ignore some import warnings. The E402 in build.py could probably
also be solved by replacing the if-statement above it by a
try-except-ImportError (why not this?).

Add spaces before class definitions, but also space out the rest of the
lines to separate each main step in `setuptools_entry`.

Relax the line length restriction to 90 instead of excessive tweaking.