Remove half-hearted attempt to create __init__.py
This never worked in the first place, because it created the package
directory if missing but not the actual __init__.py, which would have
crashed the following open() anyway.

Since we are now iterating over import packages instead of hoping that
there is a directory with the name of the distribution, we should be
safe. A regular import package should have a top-level __init__.py
in both its source and build directories. The end user wants this
__init__.py anyway, since it can contain a VERSION CHECK block for
local use of katversion at the minimum.