Put replacement __version__ assignment in place of the marked block
This fixes an issue where the module refers to `__version__` at import
time. It might work with an editable install (because it is defined
early enough), but the katversion baking process removes the assignments
to `__version__` and replaces them with an assignment at the end of
`__init__.py`, which may be too late.

Instead, when removing a BEGIN VERSION CHECK/END VERSION CHECK block,
insert the baked assignment into the same place in the file, so that
`__version__` is available at the same point it was in the original.

If an sdist built with this new version is later installed using the
older version of katversion, it'll add another `__version__` assignment
to the end of the module, but that should be harmless.